Update: Fixes.

+++ get-xpipe.sh
@@ -1,35 +1,47 @@
 #!/usr/bin/env bash
 
 release_url() {
-  echo "https://github.com/xpipe-io/xpipe/releases/latest/download"
+  local repo="$1"
+  local version="$2"
+  if [[ -z "$version" ]] ; then
+    echo "$repo/releases/latest/download"
+  else
+    echo "$repo/releases/download/$version"
+  fi
 }
 
 get_file_ending() {
   local uname_str="$(uname -s)"
   case "$uname_str" in
-    Linux)
-      if [ -f "/etc/debian_version" ]; then
-        echo "deb"
-      else
-        echo "rpm"
-      fi
-      ;;
-    Darwin)
-      echo "pkg"
-      ;;
-    *)
-      exit 1
+  Linux)
+    if [ -f "/etc/debian_version" ]; then
+      echo "deb"
+    else
+      echo "rpm"
+    fi
+    ;;
+  Darwin)
+    echo "pkg"
+    ;;
+  *)
+    exit 1
+    ;;
   esac
 }
 
 download_release_from_repo() {
   local os_info="$1"
   local tmpdir="$2"
-  local ending=$(get_file_ending)
+  local repo="$3"
+  local version="$4"
 
-  local filename="xpipe-installer-$os_info-x86_64.$ending"
+  local ending=$(get_file_ending)
+  local arch="$(uname -m)"
+  local release_url=$(release_url "$repo" "$version")
+
+  local filename="xpipe-installer-$os_info-$arch.$ending"
   local download_file="$tmpdir/$filename"
-  local archive_url="$(release_url)/$filename"
+  local archive_url="$release_url/$filename"
 
   info "Downloading file $archive_url"
   curl --progress-bar --show-error --location --fail "$archive_url" --output "$download_file" --write-out "$download_file"
@@ -67,17 +79,18 @@ parse_os_name() {
   local arch="$(uname -m)"
 
   case "$uname_str" in
-    Linux)
-      echo "linux"
-      ;;
-    FreeBSD)
-      echo "linux"
-      ;;
-    Darwin)
-      echo "macos"
-      ;;
-    *)
-      return 1
+  Linux)
+    echo "linux"
+    ;;
+  FreeBSD)
+    echo "linux"
+    ;;
+  Darwin)
+    echo "macos"
+    ;;
+  *)
+    return 1
+    ;;
   esac
   return 0
 }
@@ -85,18 +98,25 @@ parse_os_name() {
 uninstall() {
   local uname_str="$(uname -s)"
   case "$uname_str" in
-    Linux)
+  Linux)
+    if [ -d "/opt/xpipe" ]; then
+      info "Uninstalling previous version"
       if [ -f "/etc/debian_version" ]; then
         DEBIAN_FRONTEND=noninteractive sudo apt-get remove -qy xpipe
       else
         sudo rpm -e xpipe
       fi
-      ;;
-    Darwin)
-      sudo /Applications/X-Pipe.app/Contents/Resources/scripts/uninstall.sh
-      ;;
-    *)
-      exit 1
+    fi
+    ;;
+  Darwin)
+    if [ -d "/Applications/XPipe.app" ]; then
+      info "Uninstalling previous version"
+      sudo /Applications/XPipe.app/Contents/Resources/scripts/uninstall.sh
+    fi
+    ;;
+  *)
+    exit 1
+    ;;
   esac
 }
 
@@ -105,18 +125,19 @@ install() {
   local file="$1"
 
   case "$uname_str" in
-    Linux)
-      if [ -f "/etc/debian_version" ]; then
-        DEBIAN_FRONTEND=noninteractive sudo apt-get install -qy "$file"
-      else
-        sudo rpm -i "$file"
-      fi
-      ;;
-    Darwin)
-      sudo installer -verboseR -allowUntrusted -pkg "$file" -target /
-      ;;
-    *)
-      exit 1
+  Linux)
+    if [ -f "/etc/debian_version" ]; then
+      DEBIAN_FRONTEND=noninteractive sudo apt-get install -qy "$file"
+    else
+      sudo rpm -i "$file"
+    fi
+    ;;
+  Darwin)
+    sudo installer -verboseR -allowUntrusted -pkg "$file" -target /
+    ;;
+  *)
+    exit 1
+    ;;
   esac
 }
 
@@ -135,40 +156,62 @@ download_release() {
 
   # store the downloaded archive in a temporary directory
   local download_dir="$(mktemp -d)"
-  download_release_from_repo "$os_info" "$download_dir"
+  local repo="$1"
+  local version="$2"
+  download_release_from_repo "$os_info" "$download_dir" "$repo" "$version"
 }
 
 check_architecture() {
   local arch="$1"
   case "$arch" in
-    x86_64)
+  x86_64)
+    return 0
+    ;;
+  amd64)
+    return 0
+    ;;
+  arm64)
+    if [ "$(uname -s)" = "Darwin" ]; then
       return 0
-      ;;
-    amd64)
-      return 0
-      ;;
-    arm64)
-      if [ "$(uname -s)" = "Darwin" ]; then
-        return 0
-      fi
-      ;;
+    fi
+    ;;
   esac
 
-  error "Sorry! X-Pipe currently only provides pre-built binaries for x86_64 architectures."
+  error "Sorry! XPipe currently does not provide your processor architecture."
   return 1
 }
-
 
 # return if sourced (for testing the functions above)
 return 0 2>/dev/null
 
 check_architecture "$(uname -m)" || exit 1
 
-download_archive="$(download_release; exit "$?")"
+repo="https://github.com/xpipe-io/xpipe"
+version=
+while getopts 'sv:' OPTION; do
+  case "$OPTION" in
+    s)
+      repo="https://github.com/xpipe-io/xpipe_staging"
+      ;;
+
+    v)
+      version="$OPTARG"
+      ;;
+
+    ?)
+      echo "Usage: $(basename $0) [-s] [-v <version>]"
+      exit 1
+      ;;
+  esac
+done
+
+download_archive="$(
+  download_release "$repo" "$version"
+  exit "$?"
+)"
 exit_status="$?"
-if [ "$exit_status" != 0 ]
-then
-  error "Could not download X-Pipe release."
+if [ "$exit_status" != 0 ]; then
+  error "Could not download XPipe release."
   exit "$exit_status"
 fi
